Added Backgammon Project via upload
Our goal was to design and develop a turn-based app that allows two users to play backgammon against each other. It was developed with my teammate in Java and Java Swing to create graphics. The UI consists of a board panel which checkers move around, information panel which displays prompts to users, and a command panel which allows players to enter text commands.